tests reorganized in packages
